requirements.txt
streamlit
PyPDF2
openai
python-dotenv